+++ packages/axiom-components/src/Dialog/Dialog.js
@@ -24,14 +24,15 @@ export default class Dialog extends Component {
     ]),
     /** Theme applied to the overlay */
     overlayTheme: PropTypes.oneOf(['day', 'night']),
-    /** Maximum size of the Dialog */
-    size: PropTypes.oneOf(['small', 'medium', 'large']),
+    /** Toggle if the Dialog should be closed by pressing Esc */
+    shouldCloseOnEsc: PropTypes.bool,
     /** Theme of the dialog */
     theme: PropTypes.oneOf(['day', 'night']),
+    /** Custom width for Dialog */
+    width: PropTypes.string,
   };
 
   static defaultProps = {
-    size: 'small',
     theme: 'day',
   };
 
@@ -48,18 +49,17 @@ export default class Dialog extends Component {
   render() {
     const {
       children,
-      size,
       fullscreen,
       onRequestClose,
       overlayShade,
       overlayTheme,
       theme,
+      width,
       ...rest
     } = this.props;
 
     const classes = classnames('ax-dialog', {
       'ax-dialog--fullscreen': fullscreen,
-      [`ax-dialog--${size}`]: !fullscreen && size,
     });
 
     return (
@@ -67,7 +67,7 @@ export default class Dialog extends Component {
           onOverlayClick={ onRequestClose }
           overlayShade={ overlayShade }
           overlayTheme={ overlayTheme }>
-        <Base className={ classes } theme={ theme }>
+        <Base className={ classes } style={ { width } } theme={ theme }>
           { children }
         </Base>
       </Modal>