+++ packages/axiom-components/src/Position/Position.js
@@ -15,6 +15,13 @@ import './Position.css';
 export default class Position extends Component {
   static propTypes = {
     /**
+     * Allows the boundary element for the positioning to be set.
+     */
+    boundariesElement: PropTypes.oneOfType([
+      PropTypes.object,
+      PropTypes.oneOf(['scrollParent', 'viewport', 'window']),
+    ]),
+    /**
      * Children inside Position this should contain all of and
      * only PositionSource and PositionTarget!
      */
@@ -31,10 +38,16 @@ export default class Position extends Component {
     /** Controls the starting offset of the content */
     offset: PropTypes.oneOf(['start', 'middle', 'end']),
     /**
+     * When provided a mask will be placed behind PositionSource, where this
+     * function is called when clicked.
+     */
+    onMaskClick: PropTypes.func,
+    /**
      * Optional handler that is called, with the new position, when PositionSource
      * has been positioned.
      */
     onPositionChange: PropTypes.func,
+
     /**
      * Controls the starting position around PositionTarget in which the
      * PositionSource will attempt to be placed. If that position is not available
@@ -47,11 +60,12 @@ export default class Position extends Component {
   };
 
   static defaultProps = {
+    boundariesElement: 'viewport',
     enabled: true,
     flip: 'clockwise',
     offset: 'middle',
     position: 'top',
-    showArrow:  true,
+    showArrow:  false,
   };
 
   constructor(props) {
@@ -89,7 +103,7 @@ export default class Position extends Component {
   }
 
   createPopper() {
-    const { flip, showArrow } = this.props;
+    const { boundariesElement, flip, showArrow } = this.props;
     const { placement } = this.state;
 
     return new popperJS(this._target, this._content, {
@@ -104,8 +118,15 @@ export default class Position extends Component {
         flip: {
           behavior: getPlacementFlipOrder(placement, flip),
         },
-        inner: { enabled: false },
-        offset: { enabled: false },
+        inner: {
+          enabled: false,
+        },
+        offset: {
+          enabled: false,
+        },
+        preventOverflow: {
+          boundariesElement,
+        },
       },
     });
   }
@@ -149,7 +170,7 @@ export default class Position extends Component {
   }
 
   render() {
-    const { children, enabled, isVisible, showArrow, ...rest } = this.props;
+    const { children, enabled, isVisible, onMaskClick, showArrow, ...rest } = this.props;
     const { placement } = this.state;
     const [ position ] = placementToPosition(placement);
 
@@ -171,6 +192,10 @@ export default class Position extends Component {
       enabled && isVisible ? (
         <Portal { ...props } key="portal">
           <div>
+            { onMaskClick && (
+              <div className="ax-position__mask" onClick={ onMaskClick } />
+            ) }
+
             <div className={ classes } ref={ (el) => this._content = el }>
               {
                 cloneElement(findComponent(children, PositionSourceRef), {