+++ site/components/Documentation/Resources/Components/Form.js
@@ -28,7 +28,7 @@ export default class Documentation extends Component {
               <TextInput
                   onChange={ (setValue, getValue, event) => setValue('TextInput', 'value', event.target.value) }
                   placeholder="Write in me"
-                  size="medium" />
+                  size="medium"/>
             </DocumentationShowCase>
           </GridCell>
 
@@ -46,6 +46,16 @@ export default class Documentation extends Component {
         </Grid>
 
         <Grid>
+          <GridCell>
+            <DocumentationShowCase centered>
+              <TextInput
+                  label="Lorem Ipsum"
+                  onChange={ (setValue, getValue, event) => setValue('TextInput', 'value', event.target.value) }
+                  placeholder="Write in me"
+                  size="medium"
+                  usageHint="This is a usage hint" />
+            </DocumentationShowCase>
+          </GridCell>
           <GridCell>
             <DocumentationShowCase centered>
               <InlineValidation message="There seems to be something wrong with your input">
@@ -67,7 +77,10 @@ export default class Documentation extends Component {
         ] } />
 
         <DocumentationShowCase>
-          <TextArea placeholder="Write in me" />
+          <TextArea
+              label="Lorem ipsum"
+              placeholder="Write in me"
+              usageHint="This is a usage hint" />
         </DocumentationShowCase>
 
         <DocumentationApi components={ [