+++ packages/axiom-components/CHANGELOG.md
@@ -3,7 +3,141 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-      <a name="2.10.0"></a>
+      <a name="4.3.0"></a>
+# [4.3.0](https://github.com/BrandwatchLtd/axiom/compare/@brandwatch/axiom-components@4.2.1...@brandwatch/axiom-components@4.3.0) (2018-07-11)
+
+
+### Features
+
+* **Slider:** Added slider component ([7874a1c](https://github.com/BrandwatchLtd/axiom/commit/7874a1c))
+
+
+
+
+      <a name="4.2.1"></a>
+## [4.2.1](https://github.com/BrandwatchLtd/axiom/compare/@brandwatch/axiom-components@4.2.0...@brandwatch/axiom-components@4.2.1) (2018-07-11)
+
+
+### Bug Fixes
+
+* **DropdownTarget:** Ensures callback property is always fired ([fc2a8f2](https://github.com/BrandwatchLtd/axiom/commit/fc2a8f2))
+
+
+
+
+    <a name="4.2.0"></a>
+# [4.2.0](https://github.com/BrandwatchLtd/axiom/compare/@brandwatch/axiom-components@4.1.0...@brandwatch/axiom-components@4.2.0) (2018-07-06)
+
+
+### Bug Fixes
+
+* **ChartContent:** Add mask to chart dropdowns ([55db233](https://github.com/BrandwatchLtd/axiom/commit/55db233))
+* **Position:** Add onMaskClick prop back to Position ([0ad5881](https://github.com/BrandwatchLtd/axiom/commit/0ad5881))
+
+
+### Features
+
+* **AlertDropdown:** Added withMask property ([f9e05c7](https://github.com/BrandwatchLtd/axiom/commit/f9e05c7))
+* **Dropdown:** Added withMask property ([bcab6ed](https://github.com/BrandwatchLtd/axiom/commit/bcab6ed))
+
+
+
+
+<a name="4.1.0"></a>
+# [4.1.0](https://github.com/BrandwatchLtd/axiom/compare/@brandwatch/axiom-components@4.0.0...@brandwatch/axiom-components@4.1.0) (2018-07-05)
+
+
+### Features
+
+* **TextArea:** add ability to specify usage hints ([e26d8a8](https://github.com/BrandwatchLtd/axiom/commit/e26d8a8))
+* **TextInput:** add ability to specify usage hints ([f26c95f](https://github.com/BrandwatchLtd/axiom/commit/f26c95f))
+* **UsageHint:** added a usage hint component ([4ff974c](https://github.com/BrandwatchLtd/axiom/commit/4ff974c))
+
+
+
+
+<a name="4.0.0"></a>
+# [4.0.0](https://github.com/BrandwatchLtd/axiom/compare/@brandwatch/axiom-components@3.1.1...@brandwatch/axiom-components@4.0.0) (2018-07-05)
+
+
+### Features
+
+* **Dropdown:** Set show arrow to true in Dropdown and Position component ([27b06b7](https://github.com/BrandwatchLtd/axiom/commit/27b06b7))
+
+
+### BREAKING CHANGES
+
+* **Dropdown:** arrows will now be hidden be default
+
+
+
+
+<a name="3.1.1"></a>
+## [3.1.1](https://github.com/BrandwatchLtd/axiom/compare/@brandwatch/axiom-components@3.1.0...@brandwatch/axiom-components@3.1.1) (2018-07-05)
+
+
+
+
+**Note:** Version bump only for package @brandwatch/axiom-components
+
+<a name="3.1.0"></a>
+# [3.1.0](https://github.com/HHogg/axiom/compare/@brandwatch/axiom-components@3.0.0...@brandwatch/axiom-components@3.1.0) (2018-07-03)
+
+
+### Features
+
+* **Modal:** Close Modal by pressing Esc ([7162446](https://github.com/HHogg/axiom/commit/7162446))
+
+
+
+
+<a name="3.0.0"></a>
+# [3.0.0](https://github.com/HHogg/axiom/compare/@brandwatch/axiom-components@2.12.0...@brandwatch/axiom-components@3.0.0) (2018-07-03)
+
+
+### Features
+
+* **Dialog:** Added width as a prop to Dialog ([dbc60a8](https://github.com/HHogg/axiom/commit/dbc60a8))
+
+
+### BREAKING CHANGES
+
+* **Dialog:** Dialog components size property has been removed. See docs for use with new width property. To retain the same widths as before set the widths like below
+size="small" -> width="20rem"
+size="medium" -> width="37.5rem"
+size="large" -> width="50rem"
+
+
+
+
+<a name="2.12.0"></a>
+# [2.12.0](https://github.com/HHogg/axiom/compare/@brandwatch/axiom-components@2.11.0...@brandwatch/axiom-components@2.12.0) (2018-07-03)
+
+
+### Features
+
+* **StatusBadge:** Added new StatusBadge component ([a04bcbc](https://github.com/HHogg/axiom/commit/a04bcbc))
+
+
+
+
+<a name="2.11.0"></a>
+# [2.11.0](https://github.com/thatguynamedandy/axiom/compare/@brandwatch/axiom-components@2.10.0...@brandwatch/axiom-components@2.11.0) (2018-07-02)
+
+
+### Bug Fixes
+
+* **Position:** Boundary element to be viewport ([1ff9680](https://github.com/thatguynamedandy/axiom/commit/1ff9680))
+
+
+### Features
+
+* **Icons:** Added social Icons ([4b7a7c6](https://github.com/thatguynamedandy/axiom/commit/4b7a7c6))
+
+
+
+
+<a name="2.10.0"></a>
 # [2.10.0](https://github.com/lpoulter/axiom/compare/@brandwatch/axiom-components@2.9.0...@brandwatch/axiom-components@2.10.0) (2018-06-29)
 
 
@@ -20,7 +154,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
-      <a name="2.9.0"></a>
+<a name="2.9.0"></a>
 # [2.9.0](https://github.com/thatguynamedandy/axiom/compare/@brandwatch/axiom-components@2.8.0...@brandwatch/axiom-components@2.9.0) (2018-06-22)
 
 
@@ -33,7 +167,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
-  <a name="2.8.0"></a>
+<a name="2.8.0"></a>
 # [2.8.0](https://github.com/HHogg/axiom/compare/@brandwatch/axiom-components@2.7.3...@brandwatch/axiom-components@2.8.0) (2018-06-14)