+++ packages/axiom-composites/CHANGELOG.md
@@ -3,6 +3,86 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.0.52"></a>
+## [0.0.52](https://github.com/BrandwatchLtd/axiom/compare/@brandwatch/axiom-composites@0.0.51...@brandwatch/axiom-composites@0.0.52) (2018-07-11)
+
+
+
+
+**Note:** Version bump only for package @brandwatch/axiom-composites
+
+<a name="0.0.51"></a>
+## [0.0.51](https://github.com/BrandwatchLtd/axiom/compare/@brandwatch/axiom-composites@0.0.50...@brandwatch/axiom-composites@0.0.51) (2018-07-11)
+
+
+
+
+**Note:** Version bump only for package @brandwatch/axiom-composites
+
+<a name="0.0.50"></a>
+## [0.0.50](https://github.com/BrandwatchLtd/axiom/compare/@brandwatch/axiom-composites@0.0.49...@brandwatch/axiom-composites@0.0.50) (2018-07-06)
+
+
+
+
+**Note:** Version bump only for package @brandwatch/axiom-composites
+
+<a name="0.0.49"></a>
+## [0.0.49](https://github.com/BrandwatchLtd/axiom/compare/@brandwatch/axiom-composites@0.0.48...@brandwatch/axiom-composites@0.0.49) (2018-07-05)
+
+
+
+
+**Note:** Version bump only for package @brandwatch/axiom-composites
+
+<a name="0.0.48"></a>
+## [0.0.48](https://github.com/BrandwatchLtd/axiom/compare/@brandwatch/axiom-composites@0.0.47...@brandwatch/axiom-composites@0.0.48) (2018-07-05)
+
+
+
+
+**Note:** Version bump only for package @brandwatch/axiom-composites
+
+<a name="0.0.47"></a>
+## [0.0.47](https://github.com/BrandwatchLtd/axiom/compare/@brandwatch/axiom-composites@0.0.46...@brandwatch/axiom-composites@0.0.47) (2018-07-05)
+
+
+
+
+**Note:** Version bump only for package @brandwatch/axiom-composites
+
+<a name="0.0.46"></a>
+## [0.0.46](https://github.com/HHogg/axiom/compare/@brandwatch/axiom-composites@0.0.45...@brandwatch/axiom-composites@0.0.46) (2018-07-03)
+
+
+
+
+**Note:** Version bump only for package @brandwatch/axiom-composites
+
+<a name="0.0.45"></a>
+## [0.0.45](https://github.com/HHogg/axiom/compare/@brandwatch/axiom-composites@0.0.44...@brandwatch/axiom-composites@0.0.45) (2018-07-03)
+
+
+
+
+**Note:** Version bump only for package @brandwatch/axiom-composites
+
+<a name="0.0.44"></a>
+## [0.0.44](https://github.com/HHogg/axiom/compare/@brandwatch/axiom-composites@0.0.43...@brandwatch/axiom-composites@0.0.44) (2018-07-03)
+
+
+
+
+**Note:** Version bump only for package @brandwatch/axiom-composites
+
+<a name="0.0.43"></a>
+## [0.0.43](https://github.com/thatguynamedandy/axiom/compare/@brandwatch/axiom-composites@0.0.42...@brandwatch/axiom-composites@0.0.43) (2018-07-02)
+
+
+
+
+**Note:** Version bump only for package @brandwatch/axiom-composites
+
 <a name="0.0.42"></a>
 ## [0.0.42](https://github.com/lpoulter/axiom/compare/@brandwatch/axiom-composites@0.0.41...@brandwatch/axiom-composites@0.0.42) (2018-06-29)